requirements.txt
# This is a minimal, stable set of requirements to resolve dependency conflicts.

chainlit==1.1.202
pydantic<2
langchain==0.1.20
langchain-core==0.1.52
langgraph==0.0.48
langchain-openai==0.1.0
openai>=1.10.0
python-dotenv==1.0.1
PyPDF2==3.0.1
fastapi<0.111.0
uvicorn
boto3
langchain-aws
langchain-community
typing-extensions

# Add other specific packages you need below this line
